refactor: reorganize & remove files
