--- src/components/workflow/workflow-button.tsx
@@ -1,79 +1,0 @@
-import { Pause, Play } from "lucide-react";
-import { useState } from "react";
-import {
-  AlertDialog,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
-import { Button } from "@/components/ui/button";
-import { playSound } from "@/lib/sounds";
-import { useWorkflowStore } from "@/stores/use-workflow";
-
-export function WorkflowButton() {
-  const { blocks, startExecution, isExecuting, stopExecution, hasEndBlock } =
-    useWorkflowStore();
-  const canExecuteWorkflow = blocks.length >= 2 && hasEndBlock();
-  const [isStopDialogOpen, setIsStopDialogOpen] = useState(false);
-
-  const handleStopExecution = () => {
-    setIsStopDialogOpen(true);
-  };
-
-  const confirmStopExecution = () => {
-    stopExecution();
-    playSound("workflowCancel");
-    setIsStopDialogOpen(false);
-  };
-
-  const handleStartExecution = () => {
-    if (canExecuteWorkflow) {
-      startExecution();
-    }
-  };
-
-  return (
-    <>
-      <Button
-        className="rounded-lg max-xl:w-full"
-        variant="outline"
-        onClick={isExecuting ? handleStopExecution : handleStartExecution}
-        disabled={!canExecuteWorkflow || blocks.length === 0}
-      >
-        {isExecuting ? (
-          <>
-            <Pause className="size-4" />
-            <span>Stop Workflow</span>
-          </>
-        ) : (
-          <>
-            <Play className="size-4" />
-            <span>Start Workflow</span>
-          </>
-        )}
-      </Button>
-
-      <AlertDialog open={isStopDialogOpen} onOpenChange={setIsStopDialogOpen}>
-        <AlertDialogContent className="gap-0 rounded-3xl p-0">
-          <AlertDialogHeader className="border-b p-6 text-left">
-            <AlertDialogTitle>Stop Workflow</AlertDialogTitle>
-            <AlertDialogDescription>
-              Hold on! You’re on a roll and making solid progress. Are you sure
-              you want to stop your productivity workflow now?
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <div className="flex items-center justify-end gap-2 px-6 py-4">
-            <Button
-              variant="outline"
-              onClick={() => setIsStopDialogOpen(false)}
-            >
-              Keep Going
-            </Button>
-            <Button onClick={confirmStopExecution}>Stop Workflow</Button>
-          </div>
-        </AlertDialogContent>
-      </AlertDialog>
-    </>
-  );
-}
